Agregado Hateoas al branch

+++ src/main/java/com/edutech/microservicios/service/ClienteService.java
@@ -21,7 +21,7 @@ public class ClienteService {
     }
 
     // Encontrar por Id
-    public Cliente findById (Integer id) {
+    public Cliente findById (long id) {
         return clienteRepo.findById(id).get();
     }
